Moved tests to integration. Removed build_requires (not needed)
