+++ conan/tools/gnu/pkgconfigdeps/pkgconfigdeps.py
@@ -13,9 +13,8 @@ class PkgConfigDeps(object):
         pc_files = {}
         # Get all the dependencies
         host_req = self._conanfile.dependencies.host
-        build_req = self._conanfile.dependencies.direct_build
         test_req = self._conanfile.dependencies.test
-        for _, dep in list(host_req.items()) + list(build_req.items()) + list(test_req.items()):
+        for _, dep in list(host_req.items()) + list(test_req.items()):
             # Require is not used at the moment, but its information could be used,
             # and will be used in Conan 2.0
             dep_name = str(dep)
